requirements.txt
streamlit==1.51.0
opencv-python-headless==4.12.0.88
deepface==0.0.96
tf-keras==2.20.1
pillow==12.0.0
numpy==2.3.5
tensorflow==2.20.0